Add AST tree and interpreter
- Dice expression to Abstract syntax tree
- Remove PositiveInteger and use Ardalis.GuardClause instead
- Add DiceExpression to Interpreter
- Add RepeatExpression
- Repeat an expression X number of times
- For example Critical hits repeats an expression and sums the results
- Add AdditionExpression as well as other operators
- Add Addition to interpreter as well as other operators
- Add NumberList expression

+++ Dice/Rollers/StandardDiceRoller.cs
@@ -24,7 +24,7 @@ namespace DMTools.Die.Rollers
 
         public int RollDice(Dice dice)
         {
-            return _random.Next(1, (int)dice.Sides + 1);
+            return _random.Next(1, dice.Sides + 1);
         }
     }
 }
